+++ Source/ACE/Network/Sequence/UIntSequence.cs
@@ -9,42 +9,60 @@ namespace ACE.Network.Sequence
     public class UIntSequence : ISequence
     {
         private uint value;
+        private uint maxValue = UInt32.MaxValue;
 
-        public UIntSequence(uint startingValue)
+        public UIntSequence(uint startingValue, uint maxValue = UInt32.MaxValue)
         {
             value = startingValue;
+            this.maxValue = maxValue;
         }
 
         /// <summary>
         /// Creates an instance without a starting value
         /// </summary>
         /// <param name="clientPrimed">Whether the value gets sent to client before first increment</param>
-        public UIntSequence(bool clientPrimed = true)
+        public UIntSequence(bool clientPrimed = true, uint maxValue = UInt32.MaxValue)
         {
+            this.maxValue = maxValue;
             if (clientPrimed)
                 value = 0;
             else
-                value = UInt32.MaxValue;
+                value = this.maxValue;
         }
 
-        public byte[] CurrentValue
+        public uint CurrentValue
         {
             get
             {
-                return BitConverter.GetBytes(value);
+                return value;
+            }
+        }
+        public uint NextValue
+        {
+            get
+            {
+                if (value == maxValue)
+                {
+                    value = 0;
+                    return value;
+                }
+                return ++value;
             }
         }
 
-        public byte[] NextValue
+        public byte[] CurrentBytes
         {
             get
             {
-                if (value == UInt32.MaxValue)
-                {
-                    value = 0;
-                    return BitConverter.GetBytes(value);
-                }
-                return BitConverter.GetBytes(++value);
+                return BitConverter.GetBytes(CurrentValue);
+            }
+        }
+
+        public byte[] NextBytes
+        {
+            get
+            {
+                return BitConverter.GetBytes(NextValue);
             }
         }
     }